removed sidebar
