+++ packages/customer-dashboard/StreetLevelOneDashboard/public/assets/sl1_logo.png
